// src/bridge/messages/message-type.ts
enum MessageType {
  APPEND_CHILD_TO_CONTAINER = 'APPEND_CHILD_TO_CONTAINER',
  APPEND_CHILD = 'APPEND_CHILD',
  APPEND_INITIAL_CHILD = 'APPEND_INITIAL_CHILD',
  CLEAR_CONTAINER = 'CLEAR_CONTAINER',
  COMMIT_TEXT_UPDATE = 'COMMIT_TEXT_UPDATE',
  COMMIT_UPDATE = 'COMMIT_UPDATE',
  CREATE_TEXT_INSTANCE = 'CREATE_TEXT_INSTANCE',
  CREATE_INSTANCE = 'CREATE_INSTANCE',
  FREE_CALLBACK = 'FREE_CALLBACK',
  HIDE_INSTANCE = 'HIDE_INSTANCE',
  HIDE_TEXT_INSTANCE = 'HIDE_TEXT_INSTANCE',
  INSERT_BEFORE = 'INSERT_BEFORE',
  INSERT_IN_CONTAINER_BEFORE = 'INSERT_IN_CONTAINER_BEFORE',
  INVOKE_CALLBACK = 'INVOKE_CALLBACK',
  REMOVE_CHILD_FROM_CONTAINER = 'REMOVE_CHILD_FROM_CONTAINER',
  REMOVE_CHILD = 'REMOVE_CHILD',
  START_APPLICATION = 'START_APPLICATION',
  STOP_APPLICATION = 'STOP_APPLICATION',
  UNHIDE_INSTANCE = 'UNHIDE_INSTANCE',
  UNHIDE_TEXT_INSTANCE = 'UNHIDE_TEXT_INSTANCE',
}

export default MessageType